+++ plugins/backend/rag-ai-backend-embeddings-aws/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @roadiehq/rag-ai-backend-embeddings-aws
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [2ac9477]
+  - @roadiehq/rag-ai-backend-retrieval-augmenter@0.3.0
+  - @roadiehq/rag-ai-backend@0.3.0
+
 ## 0.2.1
 
 ### Patch Changes